irdma: register netdev notifiers after IB device is registered
Previously, the netdev notifiers were being registered before
the IB device was fully initialized. This introduces a race
condition that can lead to a segfault.

Registering netdevice notifiers causes artificial events
to be synchronously generated for all the netdevs in the
system, and if the netdev associated with the IB device
is UP, this can cause ib_dispatch_event to be called
for an IB device that hasn't had its ops populated yet.

This becomes a race since ib_dispatch_event results in an
ib_cache_event call executing asynchronously via a
workqueue, which ultimately results in a call to query_port
that can potentially fault.

This is most easily reproducible when loading irdma some
time after idpf is already loaded and the netdev is UP.

Moving the notifier registration to the end of the init
sequence resolves the issue.

Signed-off-by: Jacob Moroni <jmoroni@google.com>