irdma: prevent CQP timeout during create_ah if the CQP is making progress
Similar to the normal CQP op path, we should avoid triggering
a CQP timeout if the CQP is making progress.

Additionally, the wait loop was refactored to eliminate the
unconditional 1 millisecond delay. Instead, the CQP is checked
in a tight loop with jiffies being used to detect the timeout.

This has shown to improve the latency distribution and eliminate
the previous floor of 1 ms, with no real downsides since this
function is already in an atomic context.

Signed-off-by: Jacob Moroni <jmoroni@google.com>