idpf: fix timeout stall due to few pkts inflight
The HW only flushes TX completions to host memory on an interrupt, or
when the internal 128-byte buffer for completions is full (corresponding
to 16 completions). Because of these constraints, there can be cases where
BQL sets the limit in a way that pkts in flight are less than 16 compl
descriptors, such scenario can lead to timeouts.

Signed-off-by: Jacob Moroni <jmoroni@google.com>