idpf: fix re handling
RE flag should be set with a space of 64 descriptors, but existing code
only does it when the first buffer happens to be a multiple of 64 in the
descriptor ring, when you have an aggresive application sending pkts
that use more that 1 descriptor you can ran into an scenario where you
never set the RE flag.

Do not use modulo operation and instead record last time re flag was set
and make sure once the space is met you immediately set the flag.

Signed-off-by: Jacob Moroni <jmoroni@google.com>